requirements.txt
asgiref==3.5.2
boto3==1.24.69
botocore==1.27.69
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.6
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-configurations==2.4
django-crispy-forms==1.14.0
django-registration==3.3
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
